Add a DependencyAPI to firrtl.options.Phase
This adds a DependencyAPI trait to augment TransformLike. This defines
three member functions that form the basis of defining dependencies:
prerequisites, dependents, and invalidates. Prerequisites define a set
of types that must run *before* this type. Dependents define other
types which depend on this type (a way of "injecting" a prerequisite
into another type). Invalidates defines a function that can be called
to filter some collection of types to remove them from a
Set, i.e., "invalidate" that a previous type executed.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>