Add DependencyManager and PhaseManager
Adds the DependencyManager class which can be used to determine a
sequence of TransformLikes that should be run given dependency API
constraints. A DependencyManager determines an ordering that results
in some target TransformLikes being run (without invalidations) given
an initial state (some other set of TransformLikes).

This follows the original suggestion of @azidar with one (critical)
modification to the algorithm: the (DFS-based) topological sort of
dependencies is seeded using a topological sort of the
invalidations (this uses the DiGraph.seededLinearize method). This
ensures that the number of repeated TransformLikes is kept to a minimum. (I am
not sure if this is actually optimal, however.)

This additionally adds some methods that generate Graphviz for the
Prerequisite/Invalidate graph.

A concrete PhaseManager is also added.

Signed-off-by: Schuyler Eldridge <schuyler.eldridge@ibm.com>